python/pyxbos/pyproject.toml
[tool.poetry]
name = "pyxbos"
version = "0.2.32"
description = "Python bindings for XBOS-flavored WAVEMQ and related services"
authors = ["Gabe Fierro <gtfierro@cs.berkeley.edu>"]

[tool.poetry.dependencies]
python = "^3.6"
grpcio = "^1.18"
googleapis-common-protos = "^1.5"
grpcio-tools = "^1.18"
#pandas = "^0.23.4"

tlslite-ng = "^0.7.5"
beautifulsoup4 = "^4.7"
aiogrpc = "^1.6"
xxhash = "^1.3"
jq = "^0.1.6"
toml = "^0.10.0"
[tool.poetry.dev-dependencies]
phue = "^1.1"
psutil = "^5.5"
requests = "^2.21"
pyaml = "^18.11"
pymortar = "^1.0.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
